# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs SWAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/planets": {
            "get": {
                "tags": [
                    "Planets"
                ],
                "summary": "Afficher la listes des planètes",
                "operationId": "24f72d72b9702ebaa8e3637d42c8b8f1",
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des films"
                    }
                }
            },
            "post": {
                "tags": [
                    "Films"
                ],
                "summary": "Crée une planète",
                "operationId": "75f1b7ea6dd8c85b6371a7afc3296198",
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des films"
                    }
                }
            }
        },
        "/api/planets/{id}": {
            "get": {
                "tags": [
                    "Planets"
                ],
                "summary": "Afficher une planète à l'aide de son id",
                "operationId": "4954c4dfd1547e0b42d7d4dd2618093e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la planète",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des films"
                    }
                }
            },
            "put": {
                "tags": [
                    "Planets"
                ],
                "summary": "Modifie une planète",
                "operationId": "9d014d8e5773db1c11fb1ca00af01b65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la planète",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des films"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Planets"
                ],
                "summary": "Supprimer une planète",
                "operationId": "5b1ac95d2e5db95291fd132ee5eb9d88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la planète",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des films"
                    }
                }
            }
        },
        "/api/species": {
            "get": {
                "tags": [
                    "Species"
                ],
                "summary": "Afficher la liste des espèces",
                "operationId": "33bc13e65a1f56579f5a6b06cc8590a9",
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des species"
                    }
                }
            },
            "post": {
                "tags": [
                    "Species"
                ],
                "summary": "Crée une espèce",
                "operationId": "9b236b196cd5f3b0968eeef5b10aa02e",
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des films"
                    }
                }
            }
        },
        "/api/species/{id}": {
            "get": {
                "tags": [
                    "Species"
                ],
                "summary": "Afficher une espèce à l'aide de son id",
                "operationId": "53652ca2529b0426871b2555b3526d30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'Espcèce",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne un species"
                    }
                }
            },
            "put": {
                "tags": [
                    "Species"
                ],
                "summary": "Modifie une espèce",
                "operationId": "bc40d711270f593a6f3ed50a201dcd4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'espèce",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des films"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Species"
                ],
                "summary": "Supprimer une espèce",
                "operationId": "747f252c1c39ca2412455caedb25624e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'espèce",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des films"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}